Core: Add babelModeV7 feature flag for improved babel config
