+++ examples/react-ts/main.ts
@@ -22,6 +22,7 @@ const config: StorybookConfig = {
     postcss: false,
     previewCsfV3: true,
     buildStoriesJson: true,
+    babelModeV7: true,
   },
 };
