+++ lib/core-server/src/presets/common-preset.ts
@@ -1,20 +1,25 @@
+import dedent from 'ts-dedent';
 import {
   getPreviewBodyTemplate,
   getPreviewHeadTemplate,
   getManagerMainTemplate,
   getPreviewMainTemplate,
   loadCustomBabelConfig,
-  babelConfig,
+  getStorybookBabelConfig,
   loadEnvs,
   Options,
 } from '@storybook/core-common';
+import { once } from '@storybook/node-logger';
 
 export const babel = async (_: unknown, options: Options) => {
   const { configDir, presets } = options;
+  if (options.features?.babelModeV7) {
+    return presets.apply('babelDefault', {}, options);
+  }
 
   return loadCustomBabelConfig(
     configDir,
-    () => presets.apply('babelDefault', babelConfig(), options) as any
+    () => presets.apply('babelDefault', getStorybookBabelConfig(), options) as any
   );
 };
